+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+-   Ensure `TopSecret.min_confidence_score` is respected
+
 ## [0.1.0] - 2025-07-23
 
-- Initial release
+-   Initial release